added totals to CLI interface.  started API interface with addition.

+++ calculator.py
@@ -1,18 +1,20 @@
-import decimal
 from decimal import Decimal
 
+
 def add(a, b):
-    return float(a) + float(b)
+    return Decimal(a) + Decimal(b)
+
 
 def subtract(a, b):
     return Decimal(a) - Decimal(b)
 
+
 def multiply(a, b):
-    return float(a) * float(b)
+    return Decimal(a) * Decimal(b)
+
 
 def divide(a, b):
     try:
         return Decimal(a) / Decimal(b)
     except ZeroDivisionError:
         return 0
-
